+++ async-magic.js
@@ -8,5 +8,6 @@ _api.wait = require('./lib/wait');
 _api.PromiseResolver = require('./lib/PromiseResolver');
 _api.parallel = require('./lib/parallel');
 _api.series = require('./lib/series');
+_api.sleep = require('./lib/sleep');
 
 module.exports = _api;