+++ CHANGES.md
@@ -1,3 +1,6 @@
+### 1.2.1 ###
+* Changed: Doc Structure
+
 ### 1.2.0 ###
 * Added: `.wait` function as alias of [Promise.all](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/all)
 * Added: `.PromiseResolver` to queue a promised function with given arguments for execution